requirements.txt
astroid==2.12.13
black==22.12.0
click==8.1.3
dill==0.3.6
discord.py==2.2.2
Flask==2.2.2
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
pdfrw==0.4
platformdirs==2.6.0
pylint==2.15.9
pypdf==3.2.0
tomli==2.0.1
tomlkit==0.11.6
Werkzeug==2.2.2
wrapt==1.14.1
